Fix for decoding error with wind speed gusts. Values are coming back as Doubles, not Int.

+++ ForeFlightProject/ForeFlightProject/Models/AirportModel.swift
@@ -76,7 +76,7 @@ struct ConditionsVisibility: Decodable {
 struct Wind: Decodable {
     let speedKts, direction, from: Double?
     let variable: Bool?
-    let gustSpeedKts: Int?
+    let gustSpeedKts: Double?
 }
 
 // MARK: - ReportForecast
